+++ .gitignore
@@ -1,3 +1,4 @@
 .idea/
 venv/
 .env
+__pycache__
