+++ main.py
@@ -1,16 +1,10 @@
 from flask import Flask, Response, request
 import requests
-
+from config import Config
 import handle_commands
 
-TOKEN = "6313275275:AAGfL-IMwcOnk7HllN9pFW7XtV8LJYNnCqU"
 
-NGROK_URL = "https://8ff5-2a01-6500-a046-f352-a0ff-cd65-3597-6b28.ngrok-free.app"
-TELEGRAM_INIT_WEBHOOK_URL = "https://api.telegram.org/bot{}/setWebhook?url={}/message".format(TOKEN, NGROK_URL)
-MESSAGE_URL = "https://api.telegram.org/bot{}/sendMessage?chat_id={}&text={}"
-
-
-requests.get(TELEGRAM_INIT_WEBHOOK_URL)
+requests.get(Config.TELEGRAM_INIT_WEBHOOK_URL)
 
 app = Flask(__name__)
